Normalize Docling OCR language eng→en before RapidOCR.
Parse-time mapping and RapidOcrOptions(lang=...) so CLI/sidecar Tesseract
defaults never reach the onnxruntime backend as unsupported 'eng'.
