+++ packages/vera-ingest-docling/src/vera_ingest_docling/__init__.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from vera_ingest.descriptors import PipelineDescriptor
 from vera_ingest.pipeline import UnknownIngestPipelineError
 
+from .languages import map_rapidocr_languages
 from .options import DoclingOptions, describe_pipeline
 from .pipeline import DoclingHybridPipeline
 
@@ -13,6 +14,7 @@ __all__ = [
     "DoclingOptions",
     "create_pipeline",
     "describe_pipeline",
+    "map_rapidocr_languages",
 ]
